chore: remove supported/unsupported/partial from Table exports
