+++ src/core/Table.tsx
@@ -1,14 +1,6 @@
 import { Table, TableRowHeader, TableHeader, TableBody } from "./Table/Table";
 import { TableRow } from "./Table/TableRow";
-import {
-  TableCell,
-  LabelCell,
-  HeaderCell,
-  CtaCell,
-  Supported,
-  Unsupported,
-  Partial,
-} from "./Table/TableCell";
+import { TableCell, LabelCell, HeaderCell, CtaCell } from "./Table/TableCell";
 
 export default {
   Root: Table,
@@ -20,7 +12,4 @@ export default {
   RowHeader: TableRowHeader,
   Body: TableBody,
   Header: TableHeader,
-  Supported,
-  Unsupported,
-  Partial,
 };
